Move VLAN and Bandwidth relevant methods into separate VLAN and Bandwidth Service classes.
