+++ core/src/main/java/net/es/oscars/pce/TranslationPCE.java
@@ -2,6 +2,8 @@ package net.es.oscars.pce;
 
 
 import lombok.extern.slf4j.Slf4j;
+import net.es.oscars.dto.pss.EthFixtureType;
+import net.es.oscars.dto.pss.EthJunctionType;
 import net.es.oscars.pss.PCEAssistant;
 import net.es.oscars.pss.PSSException;
 import net.es.oscars.resv.dao.ReservedBandwidthRepository;
@@ -11,7 +13,6 @@ import net.es.oscars.resv.ent.*;
 import net.es.oscars.topo.beans.TopoEdge;
 import net.es.oscars.topo.beans.TopoVertex;
 import net.es.oscars.topo.dao.UrnRepository;
-import net.es.oscars.topo.ent.IntRangeE;
 import net.es.oscars.topo.ent.UrnE;
 import net.es.oscars.topo.enums.DeviceModel;
 import net.es.oscars.topo.enums.Layer;
@@ -33,22 +34,13 @@ public class TranslationPCE {
     private TopoService topoService;
 
     @Autowired
-    private ReservedBandwidthRepository bwRepo;
-
-    @Autowired
-    private ReservedVlanRepository vlanRepo;
-
-    @Autowired
-    private ReservedPssResourceRepository pssResourceRepo;
-
-    @Autowired
     private UrnRepository urnRepository;
 
     @Autowired
-    private PruningService pruningService;
+    private VlanService vlanService;
 
     @Autowired
-    private VlanSelectionService vlanService;
+    private BandwidthService bwService;
 
     /**
      * Creates a ReservedVlanJunctionE given a request for ingress/egress traffic within a device.
@@ -79,7 +71,7 @@ public class TranslationPCE {
         }
 
         // Create a reserved junction with an empty set of fixtures / PSS resources
-        ReservedVlanJunctionE rsv_j = pceAssistant.createReservedJunction(urn, new HashSet<>(), new HashSet<>(),
+        ReservedVlanJunctionE rsv_j = createReservedJunction(urn, new HashSet<>(), new HashSet<>(),
                 pceAssistant.decideJunctionType(urn.getDeviceModel()));
 
         // Select a VLAN ID for this junction
@@ -89,7 +81,7 @@ public class TranslationPCE {
         }
 
         // Confirm that there is sufficient available bandwidth
-        boolean sufficientBandwidth = confirmSufficientBandwidth(req_j, reservedBandwidths);
+        boolean sufficientBandwidth = bwService.confirmSufficientBandwidth(req_j, reservedBandwidths);
         if(!sufficientBandwidth){
             return null;
         }
@@ -98,12 +90,12 @@ public class TranslationPCE {
         // and store them in a Reserved Fixture
         Set<RequestedVlanFixtureE> reqFixtures = req_j.getFixtures();
         for(RequestedVlanFixtureE reqFix : reqFixtures){
-            ReservedBandwidthE rsvBw = pceAssistant.createReservedBandwidth(reqFix.getPortUrn(), reqFix.getInMbps(),
+            ReservedBandwidthE rsvBw = createReservedBandwidth(reqFix.getPortUrn(), reqFix.getInMbps(),
                     reqFix.getEgMbps(), sched);
 
-            ReservedVlanE rsvVlan = pceAssistant.createReservedVlan(reqFix.getPortUrn(), fixVlanMap.get(reqFix), sched);
+            ReservedVlanE rsvVlan = createReservedVlan(reqFix.getPortUrn(), fixVlanMap.get(reqFix), sched);
 
-            ReservedVlanFixtureE rsvFix = pceAssistant.createReservedFixture(reqFix.getPortUrn(), new HashSet<>(),
+            ReservedVlanFixtureE rsvFix = createReservedFixture(reqFix.getPortUrn(), new HashSet<>(),
                     rsvVlan, rsvBw, pceAssistant.decideFixtureType(reqFix.getPortUrn().getDeviceModel()));
 
             // Add the fixtures to the Reserved Junction
@@ -155,16 +147,16 @@ public class TranslationPCE {
 
         // Get map of "Ingress" and "Egress" bandwidth availability
         Map<UrnE, Map<String, Integer>> availBwMap;
-        availBwMap = createBandwidthAvailabilityMap(reservedBandwidths);
+        availBwMap = bwService.createBandwidthAvailabilityMap(reservedBandwidths);
 
         // Returns a mapping from topovertices (ports) to an "Ingress"/"Egress" map of the total Ingress/Egress
         // Requested bandwidth at that port across both the azERO and the zaERO
-        Map<TopoVertex, Map<String, Integer>> requestedBandwidthMap = createRequestedBandwidthMap(azERO, zaERO, azMbps, zaMbps);
+        Map<TopoVertex, Map<String, Integer>> requestedBandwidthMap = bwService.createRequestedBandwidthMap(azERO, zaERO, azMbps, zaMbps);
 
         // Confirm that there is sufficient bandwidth to meet the request (given what has been reserved so far)
         // Palindromic EROs -- evaluate both directions at each port - traffic flows both ways
-        if(this.palindromicEros(azERO, zaERO)){
-            boolean sufficientBw = checkForSufficientBw(urnMap, azMbps, zaMbps, azERO, zaERO, availBwMap);
+        if(pceAssistant.palindromicEros(azERO, zaERO)){
+            boolean sufficientBw = bwService.checkForSufficientBw(urnMap, azMbps, zaMbps, azERO, zaERO, availBwMap);
             if(!sufficientBw)
                 throw new PCEException("Insufficient Bandwidth to meet requested pipe" + reqPipe.toString() +
                         " given previous reservations in flow");
@@ -172,9 +164,9 @@ public class TranslationPCE {
         // Non-Palindromic EROs -- evaluate the ports in each ERO just for traffic in the AZ or ZA direction
         else{
             // Consider A->Z ERO with bwAZ
-            boolean sufficientBwAZ = checkForSufficientBwUni(urnMap, azMbps, azERO, availBwMap);
+            boolean sufficientBwAZ = bwService.checkForSufficientBwUni(urnMap, azMbps, azERO, availBwMap);
             // Consider Z->A ERO with bwZA
-            boolean sufficientBwZA = checkForSufficientBwUni(urnMap, zaMbps, zaERO, availBwMap);
+            boolean sufficientBwZA = bwService.checkForSufficientBwUni(urnMap, zaMbps, zaERO, availBwMap);
             if(!sufficientBwAZ)
             {
                 throw new PCEException("Insufficient Bandwidth to meet requested A->Z pipe" + reqPipe.toString() +
@@ -203,7 +195,7 @@ public class TranslationPCE {
                 }
                 UrnE biUrn = urnMap.get(biPort.getUrn());
 
-                if(!sufficientBandwidthAtUrn(biUrn, availBwMap, azMbps, zaMbps)){
+                if(!bwService.sufficientBandwidthAtUrn(biUrn, availBwMap, azMbps, zaMbps)){
                     throw new PCEException("Insufficient Bandwidth to meet requested pipe" + reqPipe.toString() +
                             " given previous reservations in flow");
                 }
@@ -269,7 +261,7 @@ public class TranslationPCE {
 
             // Create or retrieve A Junction
             if(!junctionMap.containsKey(aVertex)){
-                aJunction = pceAssistant.createJunctionAndFixtures(aVertex, urnMap, deviceModels, reqPipeJunctions,
+                aJunction = createJunctionAndFixtures(aVertex, urnMap, deviceModels, reqPipeJunctions,
                         chosenVlanMap, sched);
                 junctionMap.put(aVertex, aJunction);
             }
@@ -278,7 +270,7 @@ public class TranslationPCE {
             }
             // Create Z Junction
             if(!junctionMap.containsKey(zVertex)){
-                zJunction = pceAssistant.createJunctionAndFixtures(zVertex, urnMap, deviceModels, reqPipeJunctions,
+                zJunction = createJunctionAndFixtures(zVertex, urnMap, deviceModels, reqPipeJunctions,
                         chosenVlanMap, sched);
                 junctionMap.put(zVertex, zJunction);
             }
@@ -289,7 +281,7 @@ public class TranslationPCE {
             DeviceModel aModel = deviceModels.get(aJunction.getDeviceUrn().getUrn());
             DeviceModel zModel = deviceModels.get(zJunction.getDeviceUrn().getUrn());
 
-            Set<ReservedBandwidthE> pipeBandwidths = pceAssistant.createReservedBandwidthForEROs(pipeEroMap.get("AZ"),
+            Set<ReservedBandwidthE> pipeBandwidths = createReservedBandwidthForEROs(pipeEroMap.get("AZ"),
                     pipeEroMap.get("ZA"), urnMap, requestedBandwidthMap, sched);
 
 
@@ -307,7 +299,7 @@ public class TranslationPCE {
             }
             // ETHERNET
             else{
-                Set<ReservedVlanE> pipeVlans = pceAssistant.createReservedVlanForEROs(pipeEroMap.get("AZ"),
+                Set<ReservedVlanE> pipeVlans = createReservedVlanForEROs(pipeEroMap.get("AZ"),
                         pipeEroMap.get("ZA"), urnMap, chosenVlanMap, sched);
 
                 ReservedEthPipeE ethPipe = ReservedEthPipeE.builder()
@@ -327,563 +319,210 @@ public class TranslationPCE {
     }
 
 
+
     /**
-     * Build a map of the available bandwidth at each URN. For each URN, there is a map of "Ingress" and "Egress"
-     * bandwidth available. Only port URNs can be found in this map.
-     * @param rsvBandwidths - A list of all bandwidth reserved so far
-     * @return A mapping of URN to Ingress/Egress bandwidth availability
+     * Given two lists of EROS in the AZ and ZA direction, a map of URNs, a map of the requested bandwidth at each URN,
+     * and the requested schedule, return a combined set of reserved bandwidth objects for the AZ and ZA paths
+     * @param az - The AZ vertices
+     * @param za - The ZA vertices
+     * @param urnMap - A mapping of URN string to URN object
+     * @param requestedBandwidthMap - A mapping of Vertex objects to "Ingress"/"Egress" requested bandwidth
+     * @param sched - THe requested schedule
+     * @return A set of all reserved bandwidth for every port (across both paths)
      */
-    private Map<UrnE, Map<String, Integer>> createBandwidthAvailabilityMap(List<ReservedBandwidthE> rsvBandwidths){
-        // Build a map, allowing us to retrieve a list of ReservedBandwidth given the associated URN
-        Map<UrnE, List<ReservedBandwidthE>> resvBwMap = pruningService.buildReservedBandwidthMap(rsvBandwidths);
+    public Set<ReservedBandwidthE> createReservedBandwidthForEROs(List<TopoVertex> az, List<TopoVertex> za,
+                                                                  Map<String, UrnE> urnMap,
+                                                                  Map<TopoVertex, Map<String, Integer>> requestedBandwidthMap,
+                                                                  ScheduleSpecificationE sched) {
+        // Combine the AZ and ZA EROs
+        Set<TopoVertex> combined = new HashSet<>(az);
+        combined.addAll(za);
 
-        // Build a map, allowing us to retrieve the available "Ingress" and "Egress" bandwidth at each associated URN
-        Map<UrnE, Map<String, Integer>> availBwMap = new HashMap<>();
-        urnRepository.findAll()
-                .stream()
-                .filter(urn -> urn.getReservableBandwidth() != null)
-                .forEach(urn -> availBwMap.put(urn, pruningService.getBwAvailabilityForUrn(urn, urn.getReservableBandwidth(), resvBwMap)));
+        // Store the reserved bandwidths
+        Set<ReservedBandwidthE> reservedBandwidths = new HashSet<>();
 
-        return availBwMap;
+        // For each vertex in the combined set, retrieve the requested Ingress/Egress bandwidth, and create a reserved
+        // bandwidth object
+        combined.stream().filter(requestedBandwidthMap::containsKey).forEach(vertex -> {
+            UrnE urn = urnMap.get(vertex.getUrn());
+            Integer reqInMbps = requestedBandwidthMap.get(vertex).get("Ingress");
+            Integer reqEgMbps = requestedBandwidthMap.get(vertex).get("Egress");
+
+            ReservedBandwidthE rsvBw = createReservedBandwidth(urn, reqInMbps, reqEgMbps, sched);
+            reservedBandwidths.add(rsvBw);
+
+        });
+
+        // Return the set
+        return reservedBandwidths;
     }
 
     /**
-     * Given a set of reserved junctions and reserved MPLS / ETHERNET pipes, extract the reserved bandwidth objects
-     * which fall within the requested schedule period and return them all together as a list
-     * @param reservedJunctions - Set of reserved ethernet junctions
-     * @param reservedMplsPipes - Set of reserved MPLS pipes
-     * @param reservedEthPipes - Set of reserved Ethernet pipes
-     * @param sched - The requested schedule (start and end date)
-     * @return List of all reserved bandwidth contained within reserved pipes and the reserved repository.
+     * Given two lists of EROS in the AZ and ZA direction, a map of URNs, a chosen vlan ID,
+     * and the requested schedule, return a combined set of reserved VLAN objects for the AZ and ZA paths
+     * @param az - The AZ vertices
+     * @param za - The ZA vertices
+     * @param urnMap - A mapping of URN string to URN object
+     * @param vlanMap - A mapping of URN object to assigned VLAN ID
+     * @param sched - THe requested schedule
+     * @return A set of all reserved VLAN objects for every port (across both paths)
      */
-    public List<ReservedBandwidthE> createReservedBandwidthList(Set<ReservedVlanJunctionE> reservedJunctions,
-                                                                 Set<ReservedMplsPipeE> reservedMplsPipes,
-                                                                 Set<ReservedEthPipeE> reservedEthPipes,
-                                                                 ScheduleSpecificationE sched){
-        // Retrieve all bandwidth reserved so far from pipes & junctions
-        List<ReservedBandwidthE> rsvBandwidths = retrieveReservedBandwidths(reservedJunctions);
-        rsvBandwidths.addAll(retrieveReservedBandwidthsFromMplsPipes(reservedMplsPipes));
-        rsvBandwidths.addAll(retrieveReservedBandwidthsFromEthPipes(reservedEthPipes));
-        rsvBandwidths.addAll(pruningService.getReservedBandwidth(sched.getNotBefore(), sched.getNotAfter()));
-
-        return rsvBandwidths;
-    }
-
-    /**
-     * Given a set of reserved junctions and reserved ethernet pipes, retrieve all reserved VLAN objects
-     * within the specified schedule period.
-     * @param reservedJunctions - Set of reserved junctions
-     * @param reservedEthPipes - Set of reserved ethernet pipes
-     * @param sched - Requested schedule
-     * @return
-     */
-    public List<ReservedVlanE> createReservedVlanList(Set<ReservedVlanJunctionE> reservedJunctions,
-                                                       Set<ReservedEthPipeE> reservedEthPipes,
-                                                       ScheduleSpecificationE sched){
-
-        // Retrieve all VLAN IDs reserved so far from junctions & pipes
-        List<ReservedVlanE> rsvVlans = retrieveReservedVlans(reservedJunctions);
-        rsvVlans.addAll(retrieveReservedVlansFromEthPipes(reservedEthPipes));
-        rsvVlans.addAll(pruningService.getReservedVlans(sched.getNotBefore(), sched.getNotAfter()));
-
-        return rsvVlans;
-    }
-
-    /**
-     * Build a map of the requested bandwidth at each port TopoVertex contained within the az and za EROs.
-     * @param azERO - The path in the AZ direction
-     * @param zaERO - The path in the ZA direction
-     * @param azMbps - The requested bandwidth in the AZ direction
-     * @param zaMbps - The requested bandwidth in the ZA direction
-     * @return A mapping from TopoVertex (ports only) to requested "Ingress" and "Egress" bandwidth
-     */
-    private Map<TopoVertex, Map<String, Integer>> createRequestedBandwidthMap(List<TopoEdge> azERO, List<TopoEdge> zaERO,
-                                                                             Integer azMbps, Integer zaMbps){
-        // Map a port node to a map of "Ingress" and "Egress" requested bandwidth values
-        Map<TopoVertex, Map<String, Integer>> requestedBandwidthMap = new HashMap<>();
-
-        // Iterate through the AZ edges, update the map for each port node found in the path
-        for(TopoEdge azEdge : azERO){
-            updateRequestedBandwidthMap(azEdge, azMbps, requestedBandwidthMap);
-        }
-
-        // Iterate through the ZA edges, update the map for each port node in the path
-        for(TopoEdge zaEdge : zaERO){
-            updateRequestedBandwidthMap(zaEdge, zaMbps, requestedBandwidthMap);
-        }
-
-        return requestedBandwidthMap;
-    }
-
-    /**
-     * Update the values in a mapping of port vertices to requested Ingress/Egress bandwidth.
-     * @param edge - An edge, with a vertex on the "A" side and on the "Z" side
-     * @param bandwidth - The requested bandwidth in one direction
-     * @param requestedBandwidthMap - The soon-to-be updated map of requested bandwidth per port
-     */
-    private void updateRequestedBandwidthMap(TopoEdge edge, Integer bandwidth,
-                                             Map<TopoVertex, Map<String, Integer>> requestedBandwidthMap){
-
-        // Retrieve the "A" side of the edge
-        TopoVertex nodeA = edge.getA();
-        // Retrieve the "Z" side of the edge
-        TopoVertex nodeZ = edge.getZ();
-
-        // Add node A to the map if it is a port and it is not already in the map
-        if(nodeA.getVertexType().equals(VertexType.PORT) && !requestedBandwidthMap.containsKey(nodeA)){
-            requestedBandwidthMap.put(nodeA, makeInitialBandwidthMap());
-        }
-        // Add node Z to the map if it is a port and it is not already in the map
-        if(nodeZ.getVertexType().equals(VertexType.PORT) && !requestedBandwidthMap.containsKey(nodeZ)){
-            requestedBandwidthMap.put(nodeZ, makeInitialBandwidthMap());
-        }
-
-        // All Cases: Add the requested bandwidth to the amount already stored in the map
-        // Case 1: portA -> portZ -- portA = egress, portZ = ingress
-        /*
-        if(nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-        {
-            Integer updatedEgress = requestedBandwidthMap.get(nodeA).get("Egress") + bandwidth;
-            requestedBandwidthMap.get(nodeA).put("Egress", updatedEgress);
-
-            Integer updatedIngress = requestedBandwidthMap.get(nodeZ).get("Ingress") + bandwidth;
-            requestedBandwidthMap.get(nodeZ).put("Ingress", updatedIngress);
-        }*/
-        // Case 2: portA -> deviceZ -- portA = ingress
-        if(nodeA.getVertexType().equals(VertexType.PORT) && !nodeZ.getVertexType().equals(VertexType.PORT))
-        {
-            Integer updatedIngress = requestedBandwidthMap.get(nodeA).get("Ingress") + bandwidth;
-            requestedBandwidthMap.get(nodeA).put("Ingress", updatedIngress);
-        }
-        // Case 3: deviceA -> portZ -- portZ = egress
-        else if(!nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-        {
-            Integer updatedEgress = requestedBandwidthMap.get(nodeZ).get("Egress") + bandwidth;
-            requestedBandwidthMap.get(nodeZ).put("Egress", updatedEgress);
-        }
-
-    }
-
-    /**
-     * Build an initial map of requested bandwidth in the Ingress and Egress directions
-     * @return A new map with 0 Ingress and Egress bandwidth.
-     */
-    private Map<String, Integer> makeInitialBandwidthMap(){
-        Map<String, Integer> initialMap = new HashMap<>();
-        initialMap.put("Ingress", 0);
-        initialMap.put("Egress", 0);
-        return initialMap;
-    }
-
-    /**
-     * Confirm that the two EROs are identical
-     * @param azERO - A path in one direction
-     * @param zaERO - A path in another direction
-     * @return True if they are identical, False otherwise.
-     */
-    private boolean palindromicEros(List<TopoEdge> azERO, List<TopoEdge> zaERO)
-    {
-        Set<TopoVertex> azVertices = new HashSet<>();
-        Set<TopoVertex> zaVertices = new HashSet<>();
-
-        for(TopoEdge azEdge : azERO)
-        {
-            azVertices.add(azEdge.getA());
-            azVertices.add(azEdge.getZ());
-        }
-
-        for(TopoEdge zaEdge : zaERO)
-        {
-            zaVertices.add(zaEdge.getA());
-            zaVertices.add(zaEdge.getZ());
-        }
-
-        if(azVertices.size() != zaVertices.size())
-            return false;
-
-        for(TopoVertex oneVert : azVertices)
-        {
-            if(!zaVertices.contains(oneVert))
-                return false;
-        }
-
-        Set<TopoVertex> azPorts = azVertices
-                .stream()
-                .filter(v -> v.getVertexType().equals(VertexType.PORT))
-                .collect(Collectors.toSet());
-        Set<TopoVertex> zaPorts = zaVertices
-                .stream()
-                .filter(v -> v.getVertexType().equals(VertexType.PORT))
-                .collect(Collectors.toSet());
-
-        if(azPorts.size() != zaPorts.size())
-            return false;
-
-        // Now see if all ports are traversed in both the ingress and egress directions
-        Set<TopoVertex> ingressPorts = new HashSet<>();
-        Set<TopoVertex> egressPorts = new HashSet<>();
-
-        // Identify which ports are used as ingress vs the ones that are used as egress
-        for(TopoEdge azEdge : azERO)
-        {
-            TopoVertex nodeA = azEdge.getA();
-            TopoVertex nodeZ = azEdge.getZ();
-
-            // Case 1: portA -> portZ -- portA = egress, portZ = ingress
-            if(nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                egressPorts.add(nodeA);
-                ingressPorts.add(nodeZ);
-            }
-            // Case 2: portA -> deviceZ -- portA = ingress
-            else if(nodeA.getVertexType().equals(VertexType.PORT) && !nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                ingressPorts.add(nodeA);
-            }
-            // Case 3: deviceA -> portZ -- portZ = egress
-            else if(!nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                egressPorts.add(nodeZ);
-            }
-        }
-
-        // repeat above for Z->A
-        for(TopoEdge zaEdge : zaERO)
-        {
-            TopoVertex nodeA = zaEdge.getA();
-            TopoVertex nodeZ = zaEdge.getZ();
-
-            // Case 1: portA -> portZ -- portA = egress, portZ = ingress
-            if(nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                egressPorts.add(nodeA);
-                ingressPorts.add(nodeZ);
-            }
-            // Case 2: portA -> deviceZ -- portA = ingress
-            else if(nodeA.getVertexType().equals(VertexType.PORT) && !nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                ingressPorts.add(nodeA);
-            }
-            // Case 3: deviceA -> portZ -- portZ = egress
-            else if(!nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                egressPorts.add(nodeZ);
-            }
-        }
-
-        // Now check to see if all ports used are both ingress and egress -- if so, palindromic port usage!
-        if(ingressPorts.size() != egressPorts.size())
-            return false;
-
-        if(ingressPorts.size() != azPorts.size())
-            return false;
-
-        for(TopoVertex onePort : azPorts)
-        {
-            if(!(zaPorts.contains(onePort) && ingressPorts.contains(onePort) && egressPorts.contains(onePort)))
-                return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Examine all AZ and ZA edges, confirm that the requested bandwidth can be supported given the available bandwidth.
-     * @param urnMap - A map of URN string to URN objects
-     * @param azERO - The path in the A->Z direction
-     * @param zaERO - The path in the Z->A direction
-     * @param availBwMap - A map of bandwidth availability
-     * @return True, if there is sufficient bandwidth across all edges. False, otherwise.
-     */
-    private boolean checkForSufficientBw(Map<String, UrnE> urnMap, Integer azMbps, Integer zaMbps, List<TopoEdge> azERO,
-                                         List<TopoEdge> zaERO, Map<UrnE, Map<String, Integer>> availBwMap) {
-
-        // For the AZ direction, fail the test if there is insufficient bandwidth
-        if(!sufficientBandwidthForERO(azERO, urnMap, availBwMap, azMbps, zaMbps)){
-            return false;
-        }
-
-        // For each ZA direction, fail the test if there is insufficient bandwidth
-        return sufficientBandwidthForERO(zaERO, urnMap, availBwMap, azMbps, zaMbps);
-    }
-
-    /**
-     * Examine all edges, confirm that the requested bandwidth can be supported given the available bandwidth.
-     * @param urnMap - A map of URN string to URN objects
-     * @param bwMbps - The requested bandwidth (in one direction)
-     * @param availBwMap - A map of bandwidth availability
-     * @return True, if there is sufficient bandwidth across all edges. False, otherwise.
-     */
-    private boolean checkForSufficientBwUni(Map<String, UrnE> urnMap, Integer bwMbps, List<TopoEdge> ERO,
-                                            Map<UrnE, Map<String, Integer>> availBwMap){
-        // For the AZ direction, fail the test if there is insufficient bandwidth
-        return sufficientBandwidthForEROUni(ERO, urnMap, availBwMap, bwMbps);
-    }
-
-    /**
-     * Given a particular list of edges, iterate and confirm that there is sufficient bandwidth available in both directions
-     * @param ERO - A series of edges
-     * @param urnMap - A mapping of URN strings to URN objects
-     * @param availBwMap - A mapping of URN objects to lists of available bandwidth for that URN
-     * @param azMbps - The bandwidth in the AZ direction
-     * @param zaMbps - The bandwidth the ZA direction
-     * @return True, if the segment can support the requested bandwidth. False, otherwise.
-     */
-    private boolean sufficientBandwidthForERO(List<TopoEdge> ERO, Map<String, UrnE> urnMap,
-                                              Map<UrnE, Map<String, Integer>> availBwMap, Integer azMbps,
-                                              Integer zaMbps){
-        // For each edge in that list
-        for(TopoEdge edge : ERO){
-            // Retrieve the URNs
-            String urnStringA = edge.getA().getUrn();
-            String urnStringZ = edge.getZ().getUrn();
-            if(!urnMap.containsKey(urnStringA) || !urnMap.containsKey(urnStringZ)){
-                return false;
-            }
-            UrnE urnA = urnMap.get(urnStringA);
-            UrnE urnZ = urnMap.get(urnStringZ);
-
-            // If URN A has reservable bandwidth, confirm that there is enough available
-            if(urnA.getReservableBandwidth() != null){
-                if(!sufficientBandwidthAtUrn(urnA, availBwMap, azMbps, zaMbps)){
-                    return false;
-                }
-            }
-
-            // If URN Z has reservable bandwidth, confirm that there is enough available
-            if(urnZ.getReservableBandwidth() != null){
-                if(!sufficientBandwidthAtUrn(urnZ, availBwMap, azMbps, zaMbps)){
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-
-    /**
-     * Given a list of edges, confirm that there is sufficient bandwidth available in the one direction
-     * @param ERO - A series of edges
-     * @param urnMap - A mapping of URN strings to URN objects
-     * @param availBwMap - A mapping of URN objects to lists of reserved bandwidth for that URN
-     * @param bwMbps - The bandwidth in the specified direction
-     * @return True, if the segment can support the requested bandwidth. False, otherwise.
-     */
-    private boolean sufficientBandwidthForEROUni(List<TopoEdge> ERO, Map<String, UrnE> urnMap,
-                                                 Map<UrnE, Map<String, Integer>> availBwMap, Integer bwMbps)
-    {
-        // For each edge in that list
-        for(TopoEdge edge : ERO)
-        {
-            Map<UrnE, Boolean> urnIngressDirectionMap = new HashMap<>();
-            TopoVertex nodeA = edge.getA();
-            TopoVertex nodeZ = edge.getZ();
-            String urnStringA = nodeA.getUrn();
-            String urnStringZ = nodeZ.getUrn();
-
-            if(!urnMap.containsKey(urnStringA) || !urnMap.containsKey(urnStringZ))
-            {
-                return false;
-            }
-
-            UrnE urnA = urnMap.get(urnStringA);
-            UrnE urnZ = urnMap.get(urnStringZ);
-
-            // From Port to Device -- Consider Ingress direction for this Port
-            if(nodeA.getVertexType().equals(VertexType.PORT) && !nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                urnIngressDirectionMap.put(urnA, true);
-            }
-            // From Device to Port -- Consider Egress direction for this Port
-            else if(!nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                urnIngressDirectionMap.put(urnZ, false);
-            }
-            // From Port to Port -- Consider Egress for portA, and Ingress for portZ
-            else if(nodeA.getVertexType().equals(VertexType.PORT) && nodeZ.getVertexType().equals(VertexType.PORT))
-            {
-                urnIngressDirectionMap.put(urnA, false);
-                urnIngressDirectionMap.put(urnZ, true);
-            }
-
-            // If URN A has reservable bandwidth, confirm that there is enough available
-            if(urnA.getReservableBandwidth() != null)
-            {
-                boolean ingressDirection = urnIngressDirectionMap.get(urnA);
-
-                if(!sufficientBandwidthAtUrnUni(urnA, availBwMap, bwMbps, ingressDirection))
-                {
-                    return false;
-                }
-            }
-
-            // If URN Z has reservable bandwidth, confirm that there is enough available
-            if(urnZ.getReservableBandwidth() != null)
-            {
-                boolean ingressDirection = urnIngressDirectionMap.get(urnZ);
-
-                if(!sufficientBandwidthAtUrnUni(urnZ, availBwMap, bwMbps, ingressDirection))
-                {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-
-    /**
-     * Given a specific URN, determine if there is enough bandwidth available to support the requested bandwidth
-     * @param urn - The URN
-     * @param availBwMap - Map of URNs to available Bandwidth
-     * @param inMbps - Requested ingress Mbps
-     * @param egMbps - Requested egress Mbps
-     * @return True, if there is enough available bandwidth at the URN. False, otherwise
-     */
-    private boolean sufficientBandwidthAtUrn(UrnE urn, Map<UrnE, Map<String, Integer>> availBwMap,
-                                             Integer inMbps, Integer egMbps){
-        Map<String, Integer> bwAvail = availBwMap.get(urn);
-        if(bwAvail.get("Ingress") < inMbps || bwAvail.get("Egress") < egMbps){
-            log.error("Insufficient Bandwidth at " + urn.toString() + ". Requested: " +
-                    inMbps + " In and " + egMbps + " Out. Available: " + bwAvail.get("Ingress") +
-                    " In and " + bwAvail.get("Egress") + " Out.");
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Given a specific URN, determine if there is enough bandwidth available to support the requested bandwidth in a specific (ingress/egress) direction
-     * @param urn - The URN
-     * @param availBwMap - Map of URNs to Available Bandwidth
-     * @param bwMbps - Requested Mbps
-     * @param ingressDirection - True if pruning is done based upon port ingress b/w, false if done by egress b/w
-     * @return True, if there is enough available bandwidth at the URN. False, otherwise
-     */
-    private boolean sufficientBandwidthAtUrnUni(UrnE urn, Map<UrnE, Map<String, Integer>> availBwMap,
-                                                Integer bwMbps, boolean ingressDirection){
-        Map<String, Integer> bwAvail = availBwMap.get(urn);
-
-        Integer unidirectionalBW;
-        String direction = "";
-
-        if(ingressDirection)
-        {
-            unidirectionalBW = bwAvail.get("Ingress");
-            direction = " In.";
-        }
-        else
-        {
-            unidirectionalBW = bwAvail.get("Egress");
-            direction = " Out.";
-        }
-
-        if(unidirectionalBW < bwMbps)
-        {
-            log.error("Insufficient Bandwidth at " + urn.toString() + ". Requested: " + bwMbps + direction);
-            return false;
-        }
-
-        return true;
-    }
-
-
-    /**
-     * Confirm that a requested VLAN junction supports the requested bandwidth. Checks each fixture of the junction.
-     * @param req_j - The requested junction.
-     * @param reservedBandwidths - The reserved bandwidth.
-     * @return True, if there is enough bandwidth at every fixture. False, otherwise.
-     */
-    public boolean confirmSufficientBandwidth(RequestedVlanJunctionE req_j, List<ReservedBandwidthE> reservedBandwidths){
-
-        // All requested fixtures on this junction
-        Set<RequestedVlanFixtureE> reqFixtures = req_j.getFixtures();
-
-
-        // Get map of "Ingress" and "Egress" bandwidth availability
-        Map<UrnE, Map<String, Integer>> availBwMap = createBandwidthAvailabilityMap(reservedBandwidths);
-
-        // For each requested fixture,
-        for(RequestedVlanFixtureE reqFix: reqFixtures){
-            // Confirum that there is enough available bandwidth at that URN
-            if(!sufficientBandwidthAtUrn(reqFix.getPortUrn(), availBwMap, reqFix.getInMbps(), reqFix.getEgMbps())){
-                return false;
-            }
-        }
-        return true;
-    }
-
-
-    /**
-     * Retrieve all reserved bandwidths from a set of reserved junctions.
-     * @param junctions - Set of reserved junctions.
-     * @return A list of all bandwidth reserved at those junctions.
-     */
-    public List<ReservedBandwidthE> retrieveReservedBandwidths(Set<ReservedVlanJunctionE> junctions){
-        return junctions
-                .stream()
-                .map(ReservedVlanJunctionE::getFixtures)
-                .flatMap(Collection::stream)
-                .map(ReservedVlanFixtureE::getReservedBandwidth)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Retrieve all reserved VLAN IDs from a set of reserved junctions
-     * @param junctions - Set of reserved junctions.
-     * @return A list of all VLAN IDs reserved at those junctions.
-     */
-    public List<ReservedVlanE> retrieveReservedVlans(Set<ReservedVlanJunctionE> junctions){
-        return junctions
-                .stream()
-                .map(ReservedVlanJunctionE::getFixtures)
-                .flatMap(Collection::stream)
-                .map(ReservedVlanFixtureE::getReservedVlan)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Retrieve all Reserved Bandwidth from a set of reserved MPLS pipes.
-     * @param reservedPipes - Set of reserved pipes
-     * @return A list of all reserved bandwidth within the set of reserved MPLS pipes.
-     */
-    public List<ReservedBandwidthE> retrieveReservedBandwidthsFromMplsPipes(Set<ReservedMplsPipeE> reservedPipes) {
-        return reservedPipes
-                .stream()
-                .map(ReservedMplsPipeE::getReservedBandwidths)
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Retrieve all Reserved Bandwidth from a set of reserved Ethernet pipes.
-     * @param reservedPipes - Set of reserved pipes
-     * @return A list of all reserved bandwidth within the set of reserved Ethernet pipes.
-     */
-    public List<ReservedBandwidthE> retrieveReservedBandwidthsFromEthPipes(Set<ReservedEthPipeE> reservedPipes) {
-        return reservedPipes
-                .stream()
-                .map(ReservedEthPipeE::getReservedBandwidths)
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
-    }
-
-
-
-    /**
-     * Retrieve all Reserved VLAN IDs from a set of reserved pipes.
-     * @param reservedPipes - Set of reserved pipes
-     * @return A list of all reserved VLAN IDs within the set of reserved pipes.
-     */
-    public List<ReservedVlanE> retrieveReservedVlansFromEthPipes(Set<ReservedEthPipeE> reservedPipes) {
-        List<ReservedVlanE> reservedVlans = new ArrayList<>();
-        Set<ReservedVlanJunctionE> junctions = new HashSet<>();
-        for(ReservedEthPipeE pipe : reservedPipes){
-            junctions.add(pipe.getAJunction());
-            junctions.add(pipe.getZJunction());
-            reservedVlans.addAll(pipe.getReservedVlans());
-        }
-        reservedVlans.addAll(retrieveReservedVlans(junctions));
+    public Set<ReservedVlanE> createReservedVlanForEROs(List<TopoVertex> az, List<TopoVertex> za,
+                                                        Map<String, UrnE> urnMap, Map<UrnE, Integer> vlanMap,
+                                                        ScheduleSpecificationE sched) {
+        // Combine the AZ and ZA EROs
+        Set<TopoVertex> combined = new HashSet<>(az);
+        combined.addAll(za);
+
+        // Store the reserved bandwidths
+        Set<ReservedVlanE> reservedVlans = new HashSet<>();
+
+        // For each vertex in the combined set, retrieve the requested Ingress/Egress bandwidth, and create a reserved
+        // bandwidth object
+        combined.stream().filter(vertex -> vertex.getVertexType().equals(VertexType.PORT)).forEach(vertex -> {
+            UrnE urn = urnMap.get(vertex.getUrn());
+
+            ReservedVlanE rsvVlan = createReservedVlan(urn, vlanMap.get(urn), sched);
+            reservedVlans.add(rsvVlan);
+
+        });
+
+        // Return the set
         return reservedVlans;
     }
+
+    /**
+     * Create a Junction and it's associated fixtures given the input parameters.
+     * @param device - The device vertex associated with the junction
+     * @param urnMap - A mapping of URN strings to URN objects
+     * @param deviceModels - A mapping of URN strings to Device Models
+     * @param requestedJunctions - A set of requested junctions - used to determine attributes of junction/fixtures
+     * @param vlanMap - Map of assigned VLANs for each fixture
+     * @param sched - The requested schedule
+     * @return A Reserved Vlan Junction with Reserved Fixtures (if contained in a matching requested junction)
+     * @throws PSSException
+     */
+    public ReservedVlanJunctionE createJunctionAndFixtures(TopoVertex device, Map<String, UrnE> urnMap,
+                                                           Map<String, DeviceModel> deviceModels,
+                                                           Set<RequestedVlanJunctionE> requestedJunctions,
+                                                           Map<UrnE, Integer> vlanMap, ScheduleSpecificationE sched)
+            throws PSSException {
+        // Get this junction's URN, Device Model, and Typing for its Fixtures
+        UrnE aUrn = urnMap.get(device.getUrn());
+        DeviceModel model = deviceModels.get(aUrn.getUrn());
+        EthFixtureType fixType = pceAssistant.decideFixtureType(model);
+
+        // Create the set of Reserved Fixtures by filtering for requested junctions that match the URN
+        log.info(requestedJunctions.stream().filter(reqJunction -> reqJunction.getDeviceUrn().equals(aUrn)).collect(Collectors.toList()).toString());
+        Set<ReservedVlanFixtureE> reservedVlanFixtures = requestedJunctions
+                .stream()
+                .filter(reqJunction -> reqJunction.getDeviceUrn().equals(aUrn))
+                .map(RequestedVlanJunctionE::getFixtures)
+                .flatMap(Collection::stream)
+                .map(reqFix -> createFixtureAndResources(reqFix.getPortUrn(), fixType,
+                        reqFix.getInMbps(), reqFix.getEgMbps(), vlanMap.get(reqFix.getPortUrn()), sched))
+                .collect(Collectors.toSet());
+
+        log.info("Reserved Fixtures at " + aUrn + ": " + reservedVlanFixtures);
+        // Return the Reserved Junction (with the set of reserved VLAN fixtures).
+        return createReservedJunction(aUrn, new HashSet<>(),
+                reservedVlanFixtures, pceAssistant.decideJunctionType(model));
+    }
+
+    /**
+     * Create a Reserved Fixtures and it's associated resources (VLAN and Bandwidth) using the input.
+     * @param portUrn - The URN of the desired fixture
+     * @param fixtureType - The typing of the desired fixture
+     * @param azMbps - The requested ingress bandwidth
+     * @param zaMbps - The requested egress bandwidth
+     * @param vlanId - The assigned VLAN ID
+     * @param sched - The requested schedule
+     * @return The reserved fixture, containing all of its reserved resources
+     */
+    public ReservedVlanFixtureE createFixtureAndResources(UrnE portUrn, EthFixtureType fixtureType, Integer azMbps,
+                                                          Integer zaMbps, Integer vlanId,
+                                                          ScheduleSpecificationE sched){
+
+
+        // Create reserved resources for Fixture
+        ReservedBandwidthE rsvBw = createReservedBandwidth(portUrn, azMbps, zaMbps, sched);
+        ReservedVlanE rsvVlan = createReservedVlan(portUrn, vlanId, sched);
+        // Create Fixture
+        return createReservedFixture(portUrn, new HashSet<>(), rsvVlan, rsvBw, fixtureType);
+    }
+
+    /**
+     * Create a reserved junction given the input
+     * @param urn - The junction's URN
+     * @param pssResources - The junction's PSS Resources
+     * @param fixtures - The junction's fixtures
+     * @param junctionType - The junction's type
+     * @return The Reserved VLAN Junction
+     */
+    public ReservedVlanJunctionE createReservedJunction(UrnE urn, Set<ReservedPssResourceE> pssResources,
+                                                        Set<ReservedVlanFixtureE> fixtures, EthJunctionType junctionType){
+        return ReservedVlanJunctionE.builder()
+                .deviceUrn(urn)
+                .reservedPssResources(pssResources)
+                .fixtures(fixtures)
+                .junctionType(junctionType)
+                .build();
+    }
+
+    /**
+     * Create a reserved fixture, given the input parameters.
+     * @param urn - The fixture's URN
+     * @param pssResources - The fixture's PSS Resources
+     * @param rsvVlan - The fixture's assigned VLAN ID
+     * @param rsvBw - The fixture's assigned bandwidth
+     * @param fixtureType - The fixture's type
+     * @return The Reserved VLAN Fixture
+     */
+    public ReservedVlanFixtureE createReservedFixture(UrnE urn, Set<ReservedPssResourceE> pssResources,
+                                                      ReservedVlanE rsvVlan, ReservedBandwidthE rsvBw,
+                                                      EthFixtureType fixtureType){
+        return ReservedVlanFixtureE.builder()
+                .ifceUrn(urn)
+                .reservedPssResources(pssResources)
+                .reservedVlan(rsvVlan)
+                .reservedBandwidth(rsvBw)
+                .fixtureType(fixtureType)
+                .build();
+    }
+
+
+    /**
+     * Create the reserved bandwidth given the input parameters.
+     * @param urn - The URN associated with this bandwidth
+     * @param inMbps - The ingress bandwidth
+     * @param egMbps - The egress bandwidth
+     * @param sched - The requested schedule
+     * @return A reserved bandwidth object
+     */
+    public ReservedBandwidthE createReservedBandwidth(UrnE urn, Integer inMbps, Integer egMbps, ScheduleSpecificationE sched){
+        return ReservedBandwidthE.builder()
+                .urn(urn)
+                .inBandwidth(inMbps)
+                .egBandwidth(egMbps)
+                .beginning(sched.getNotBefore().toInstant())
+                .ending(sched.getNotAfter().toInstant())
+                .build();
+    }
+
+    /**
+     * Create the reserved VLAN ID given the input parameters.
+     * @param urn - The URN associated with this VLAN
+     * @param vlanId - The ID value for the VLAN tag
+     * @param sched - The requested schedule
+     * @return The reserved VLAN objct
+     */
+    public ReservedVlanE createReservedVlan(UrnE urn, Integer vlanId, ScheduleSpecificationE sched){
+        return ReservedVlanE.builder()
+                .urn(urn)
+                .vlan(vlanId)
+                .beginning(sched.getNotBefore().toInstant())
+                .ending(sched.getNotAfter().toInstant())
+                .build();
+    }
+
 }
